Show warning when fails to retrieve latest Plex download URL.
Fixed #14.

+++ install_plex.sh
@@ -1,11 +1,17 @@
 #!/bin/sh -ex
 
 # Download and install Plex Media Server
-curl -L $(/retrieve_plex_download_url.py) -o plexmediaserver.deb
-dpkg -i plexmediaserver.deb
+DOWNLOAD_URL=$(/retrieve_plex_download_url.py)
+if [ $? -eq 2 ]
+then
+    echo "Skip install because it failed to retrieve latest Plex download URL."
+else
+    curl -L $DOWNLOAD_URL -o plexmediaserver.deb
+    dpkg -i plexmediaserver.deb
+    rm -f plexmediaserver.deb
+fi
 
 # Clean-up (we also don't need the PlexPass login/password after this point
 # so it's safer if we also unset them).
 unset PLEXPASS_LOGIN
 unset PLEXPASS_PASSWORD
-rm -f plexmediaserver.deb
